refactor. 출력에 해당하는 로직을 OutputView 클래스로 이동
 - view패키지의 클래스명을 대문자로 시작하도록 convention 수정
 - RacingCarApp클래스에서 출력하는 로직을 OutputView클래스로 이동
 - RacingCars에서 우승자를 String으로 변환하는 로직을 OutputView 클래스로 이동
 - RacingCars의 setCars 메소드 이름을 createCars로 변경
 - RacingCarController에서 racingResult를 인터페이스를 사용해 참조하도록 변경
