+++ src/main/java/racingcargame/controller/RacingCarController.java
@@ -1,34 +1,35 @@
 package racingcargame.controller;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import racingcargame.domain.CarStatus;
 import racingcargame.domain.MoveDecider;
 import racingcargame.domain.RacingCars;
 import racingcargame.domain.RepeatTimes;
 
-import java.util.ArrayList;
-
 public class RacingCarController {
 
-    public String winner;
+	public List<String> winners;
 
-    public ArrayList<CarStatus> run(String inputName, RepeatTimes inputRepeat) {
-        final RacingCars racingCars = new RacingCars(inputName);
-        final MoveDecider moveDecider = new MoveDecider();
+	public List<CarStatus> run(String inputName, RepeatTimes inputRepeat) {
+		final RacingCars racingCars = new RacingCars(inputName);
+		final MoveDecider moveDecider = new MoveDecider();
 
-        final ArrayList<CarStatus> racingResult = processGame(inputRepeat, racingCars, moveDecider);
-        winner = rankGame(racingCars);
-        return racingResult;
-    }
+		final List<CarStatus> racingResult = processGame(inputRepeat, racingCars, moveDecider);
+		winners = rankGame(racingCars);
+		return racingResult;
+	}
 
-    private ArrayList<CarStatus> processGame(RepeatTimes repeatTimes, RacingCars racingCars, MoveDecider moveDecider) {
-        ArrayList<CarStatus> racingResult = new ArrayList<>();
-        for (int i = 0; i < repeatTimes.getRepeatTimes(); i++) {
-            racingResult.add(racingCars.processOneRace(moveDecider));
-        }
-        return racingResult;
-    }
+	private List<CarStatus> processGame(RepeatTimes repeatTimes, RacingCars racingCars, MoveDecider moveDecider) {
+		List<CarStatus> racingResult = new ArrayList<>();
+		for (int i = 0; i < repeatTimes.getRepeatTimes(); i++) {
+			racingResult.add(racingCars.processOneRace(moveDecider));
+		}
+		return racingResult;
+	}
 
-    private String rankGame(RacingCars racingCars) {
-        return racingCars.getWinner();
-    }
+	private List<String> rankGame(RacingCars racingCars) {
+		return racingCars.getWinners();
+	}
 }
