+++ src/main/java/racingcargame/domain/CarStatus.java
@@ -5,19 +5,19 @@ import java.util.List;
 
 public class CarStatus {
 
-    private List<Car> cars;
+	private List<Car> cars;
 
-    CarStatus(List<Car> cars) {
-        this.cars = cloneCars(cars);
-    }
+	CarStatus(List<Car> cars) {
+		this.cars = cloneCars(cars);
+	}
 
-    private List<Car> cloneCars(List<Car> cars) {
-        List<Car> tempCars = new ArrayList<>();
-        cars.forEach(car -> tempCars.add(new Car(car)));
-        return tempCars;
-    }
+	private List<Car> cloneCars(List<Car> cars) {
+		List<Car> tempCars = new ArrayList<>();
+		cars.forEach(car -> tempCars.add(new Car(car)));
+		return tempCars;
+	}
 
-    public List<Car> getCars() {
-        return cars;
-    }
+	public List<Car> getCars() {
+		return cars;
+	}
 }
