+++ src/main/java/racingcargame/domain/MoveDecider.java
@@ -4,9 +4,9 @@ import java.util.Random;
 
 public class MoveDecider extends Random {
 
-    private static final int BOUND = 10;
+	private static final int BOUND = 10;
 
-    public int getNumber() {
-        return nextInt(BOUND);
-    }
+	public int getNumber() {
+		return nextInt(BOUND);
+	}
 }
